fix sentence generator bug, update DOM events
